getnearby r => c
